Queries working. FINALLY!
Took a long time but the system effectively forwards messages as it should. Right now it'll forward the exit message to safely terminate all nodes.
Haven't run valgrind on it, and something tells me there will be some surprises.
Also, the polling loop is really demanding in terms of performance...